asynchronously call send email
